fix: downgrade webpack (#243)
An updated webpack seems to be generating javascript that ends up
being invalid(?) inside the VSCode execution engine. The way this
manifests is a lack of being able to find `utils` package in the node.js
stdlib.

We should look into why this webpack can't be updated, because it does
come with some `npm audit` noisiness, but for now, this updated webpack
makes a VSCode extension that doesn't even work.