fix: address 404s in static resource references
From the Department of How Did This Ever Work, I noticed that the Add
Connection ui was unstyled and didn't seem to work quite right. Upon
digging in, I noticed that the debug console was showing 404 for the css
and js files. After reading the `Webview` API docs thoroughly and not
really seeing breakages in access control, I realized that the `cssPath`
and `jsPath` mustache variables were surrounded with `{{}}`, which is
html entity escaped by default, whereas `{{{}}}` renders the string
as-is. Marking those variables as safe fixed the issue. I am unsure how
that _ever_ worked.