fix: update influxdb version toggle
The earlier influxdb version toggle didn't completely work. It didn't
set the hidden elements to be disabled, so they were trying to validate
while hidden (and thus the validation messages couldn't figure out where
to display the error message). When submitting the data, the detection
for which version was selected was using `0` in the logic rather than
`"0"`, which made the submission always think it was using influxdbv1.
This is the fix for #232.

Additionally, I discovered that `editConn.html` had invalid markup,
with erroneous `</div>` tags. Once that worked, the VSCode html plugin
then changed the whitespace. I agree with the changes the plugin made,
so I recommend viewing this patch ignoring whitespace.

Closes #232