fix: fix connection listing (#269)
Somehow, the update to show table rows broke the treelist on the left
hand side and I didn't catch it. I'm certainly concerned about it, but I
don't currently have a way to prevent this regression outside of
`npm run clean` and manually verifying master before release.

I experienced this issue, but it was also reported as part of #266.