feat: use the official api client
This patch essentially re-writes the vscode plugin to use the official
influxdb client. It refactors the tree view to be more straight-forward,
and removes the unnecessary nesting of functionality. In the process,
this refactor made it quite straight-forward to fix a number of
outstanding issues that were otherwise difficult to address.

Closes #274
Closes #258
Closes #183
Closes #267