feat: open task for editing (#286)
This patch adds the ability to edit a task. In the tree view, find the
task to edit, right click and select "Edit task", and the task will be
opened in the editor. Edit the task (and run it, should you like), and
upon save, the contents of the file will be written back to the remote
influxdb instance.

Closes #277