fix: clean up the filesystem after editing a task (#288)
While recording a demo, I discovered an issue where the old file isn't
cleaned up. I had originally assumed that `os.tmpdir` would create a new
directory every time, but it just returns the existing one that has a
bunch of other files in it. This patch creates a random dir in that dir,
writes the file to that dir, and then cleans up the whole dir after
save.