fix: bind the event handlers (#289)
Usually, event handlers bind `this` to the event itself. It seems that
vscode seems to just not bind it at all, so `this` is `undefined`. The
issue with this is that the language client's event handlers then
reference `this` as though it's an instance of `Client`. This has never
actually worked, but I started seeing more often when I was testing the
task editing work.