fix: display the table output more correctly
This patch does a few things all related to how the table output is
shown. First, it uses an asterisk to indicate group keys. Second, it
fixes a bug where each row appeared to be its own table. Lastly, it
populates the `_result` field on the rows based on the default. It seems
that it doesn't show up in the row itself, so has to be divined from the
table head.

Closes #290