feat: allow disabling TLS verification
This patch adds support for disabling TLS verification. It adds a
property to the interface that is stored in vscode, and then adds the
plumbing to flip the bit in the UI.

Additionally, this patch introduces a `Store` singleton that allows us
to query various storage mechanisms in Code. Currently, it just
abstracts the current context global state. By doing this, we ensure
that the data model is enforced in one place (see the logic around
`isActive').

Closes #202