chore: consolidate release process
The release process used to be a two-step process, with a patch to
update the version and then a tag once that patch has merged. Now it can
all be done in one-fell swoop, with a commit directly to master.

This patch is that single script that handles all of the needed steps to
do a release. It's _maybe_ a little overly strict, but I'd rather it be
super strict than release things when they shouldn't be.