feat: add initial support for API Invoked Functions
This patch adds initial (and slightly crude) support for API Invocable
Functions. Functions can be added/edited/deleted through the InfluxDB
pane in vscode.

This functionality is currently behind a false statement, which is a
poor man's feature flag. There's a little bit of polish that this
experience needs to provide, and there's still no way to edit the
scripts.