feat: store the org id with the instance (#332)
Occasionally, the apis consumed by vsflux require the org id, rather
than the org. The way this was fixed earlier, when the org id is needed,
it is fetched. We can minimize API calls by caching that.

This patch starts by converting the "Add/Edit Instance" flow to use the
controller structure that is used in other places. When adding/editing
an instance, if the orgID isn't available, it is fetched.

Next, this patch introduces a migration API so that on startup, any
existing instances are updated. This will be useful going forward making
these types of transformations easier in the future.

Closes #326