fix: check for script capability rather than hardcoding to cloud (#334)
This patch changes the way we detect the invocable scripts feature in
vsflux. Previously, it just checked the `instance.hostNport` against a
list of cloud instances. That broke when it came to cloud instances not
publicly accessible. This patch changes the functionality to check for
the script api, which has the added benefit of future proofing the logic
as well.