chore: fix publish process to not use sudo (#347)
This patch removes the need to run `sudo ...` in the publish process.
Prior to this patch, `vsce` was installed out of band as part of the
publish process, but required `sudo` to install globally. This causes
issues later when the publish process tries to read cached files not
accessible because they are owned by `root`. Now the process is entirely
contained inside of `package.json` and we don't install anything
globally anymore.