fix: escape path separators in task/script names
Part of the (current) flow to adding/editing tasks and scripts is to
create a temporary file on the filesystem. We use the name of task or
script as the filename, but when it has characters that are path
separators, that flow breaks. This patch makes a small change to fix
that issue.

Long term, we're currently re-thinking this flow entirely, and whether
we need to write a file at all, or if there's a better flow.

Fixes #349