feat: add cli configuration as a possible data source
This patch adds support for backing the instance list using the cli
configuration rather than the vscode global state. This patch supports
create/read/update/delete operations using the `influx` command.

Upon updating the extension, a migration will attempt to see if there is
an `influx` command available. If it is, the user is prompted to choose
between vscode and the cli configuration. If the cli configuration is
chosen, all existing db-stored connections are migrated to the cli
configuration.

Fixes #295