fix: bump engine version again (#377)
_le sigh_

The error message this raises is confusing. It will say "You're using
types version X, but you need a version greater than Y", where Y is, I
guess, the current version, and has nothing to do with compatibility
with X. What it should say is "engine needs to be greater than or equal
to version X." My attempt to be conservative and not expect VSCode to be
super up-to-date (I was using 1.63.3) meant I still didn't fix it.
Updated vscode, bumped the engine version, and made a prerelease job
that allowed me to test that it'd work (but not actual publish
anything).

I will figure out how to test this better when it's not blocking a
release.