chore: tweaks to release process (#379)
The last vsflux release was an ordeal. This patch removes some of those
footguns so that it is less of an ordeal in the future.

The big change here is that we'll now publish a "pre-release" directly
from master on every merge. Users who would like to opt into pre-release
version of vsflux will then be able to get tip. I think that's pretty
neat.

Additionally, as an added benefit to having that regular build
published, the release script can check the ci status now and verify
that the publish occurred and that a release will succeed through the
pipeline should one happen.

I also fixed how the check for an up-to-date flux-lsp happens, as I'm
pretty sure it _never_ worked.