chore: fix prerelease
This patch fixes the prerelease flow, as it was less smart than I
originally expected. The version has to be bumped in `package.json`
before it can put out as a prerelease version, whereas I had assumed it
would do that automatically. Additionally, it also fixes calculating
changelogs for release notes.