chore: fix prerelease (#382)
This patch fixes the prerelease flow, as it was less smart than I
originally expected. The version has to be bumped in `package.json`
before it can put out as a prerelease version, whereas I had assumed it
would do that automatically. Because that version cannot be committed
(we don't want our CI to commit things), a release script calculates
what the prerelease version should be, as `npm version prerelease`
can't maintain that state.

Additionally, it also fixes calculating changelogs for release notes,
as that seemed to not really work correctly.