fix: small fixes to Add Connection flow
It seems that the Add Connection flow got broken a bit when adding
support for the cli backend. This patch fixes the core issue described
in #375, but also fixes a couple of issues discovered in testing that
fix, mainly that spaces weren't allowed in name or org, and that the
messages from the "Test Connection" flow weren't working.

Fixes #375