feat: select the connection when running flux (#385)
This patch removes a lot of the need around "Active" connections. It
prompts for which connection to run the flux against when F5 is hit.
This will squish a whole class of bugs.

Additionally, I removed the command for running the query outside of
this flow. This removes the need for "Active" connection entirely,
though that functionality is left for workspace state for upcoming LSP
features.