fix: support pagination in tasks and scripts
Currently, we only get the first page of tasks and scripts. That means
that if you have more than the default 20 tasks or scripts, you only get
the first page in vscode. This patch fixes that, so now vscode will just
keep fetching tasks and scripts until there are no more to fetch.

Fixes #404