fix: only add connections that can be authenticated
In the current plugin, when one tries to add a connection that either
has bad auth information or a bad network connection, the "save"
functionality has no feedback. Making a connection to the database is
important, as that's how we get the org id, which is needed for a few of
the requests we make.

This patch removes the "Test" functionality, as it is now unneeded, and
prevents adding databases with bad auth information.

Fixes #400