fix: remove accidental caching of tasks and scripts
When editing scripts, we use the representation we have fetched and held
in memory when the treeview was rendered. Unfortunately, this means that
tasks and scripts are unintentionally cached, as upon edit, we use that
version from the treeview render. This patch changes that to refetch the
object on edit, so we can get the most up-to-date representation on
edit.

Fixes #407