feat: allow changing settings via settings page (#417)
Until the introduction of the cli datasource, there wasn't really a use
for the settings page of the extension. Even then, that feature was
mostly experimental, and we didn't want to expose it. At this point, I
think it makes sense to expose that on the settings page. This patch
exposes that setting on the settings page and removes an holdover from
the previous v1 support.